scan_generation moved to core
